requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
yfinance==0.2.18
alpha-vantage==2.3.1
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
streamlit==1.25.0
joblib==1.3.2
PyYAML==6.0.1
python-dotenv==1.0.0
requests==2.31.0
pymc==5.7.2
arviz==0.15.1
# Use tensorflow-cpu to avoid long path issues and large installs on Windows
tensorflow-cpu==2.14.0
keras==2.14.0
xgboost==1.7.6
lightgbm==4.0.0
optuna==3.3.0
mlflow==2.5.0
docker==6.1.3
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
